+++ gen/templates/policy_definitions/resource.tf
@@ -11,11 +11,23 @@ resource "sdwan_{{snakeCase .Name}}_policy_definition" "example" {
       {{- if eq .Type "List"}}
         {{.TfName}} = [
           {
-            {{- range  .Attributes}}
-            {{- if and (not .ExcludeTest) (not .ExcludeExample) (not .TfOnly) (not .Value)}}
-            {{.TfName}} = {{if eq .Type "String"}}"{{end}}{{.Example}}{{if eq .Type "String"}}"{{end}}
-            {{- end}}
-            {{- end}}
+          {{- range  .Attributes}}
+          {{- if and (not .ExcludeTest) (not .ExcludeExample) (not .TfOnly) (not .Value)}}
+          {{- if eq .Type "List"}}
+            {{.TfName}} = [
+              {
+                {{- range  .Attributes}}
+                {{- if and (not .ExcludeTest) (not .ExcludeExample) (not .TfOnly) (not .Value)}}
+                {{.TfName}} = {{if eq .Type "String"}}"{{end}}{{.Example}}{{if eq .Type "String"}}"{{end}}
+                {{- end}}
+                {{- end}}
+              }
+            ]
+          {{- else}}
+          {{.TfName}} = {{if eq .Type "String"}}"{{end}}{{.Example}}{{if eq .Type "String"}}"{{end}}
+          {{- end}}
+          {{- end}}
+          {{- end}}
           }
         ]
       {{- else}}